Split tests between controller and flask-restx
