+++ tests/test_flask_restx_mongo_controller_auto_increment.py
@@ -1,0 +1,254 @@
+import enum
+
+import flask
+import flask_restx
+import pytest
+
+import layabase
+import layabase.mongo
+
+
+class EnumTest(enum.Enum):
+    Value1 = 1
+    Value2 = 2
+
+
+@pytest.fixture
+def controller() -> layabase.CRUDController:
+    class TestCollection:
+        __collection_name__ = "test"
+
+        key = layabase.mongo.Column(
+            int, is_primary_key=True, should_auto_increment=True
+        )
+        enum_field = layabase.mongo.Column(
+            EnumTest, is_nullable=False, description="Test Documentation"
+        )
+        optional_with_default = layabase.mongo.Column(str, default_value="Test value")
+
+    controller = layabase.CRUDController(TestCollection)
+    layabase.load("mongomock", [controller])
+    return controller
+
+
+@pytest.fixture
+def app(controller: layabase.CRUDController):
+    application = flask.Flask(__name__)
+    application.testing = True
+    api = flask_restx.Api(application)
+    namespace = api.namespace("Test", path="/")
+
+    controller.flask_restx.init_models(namespace)
+
+    @namespace.route("/test")
+    class TestResource(flask_restx.Resource):
+        @namespace.expect(controller.flask_restx.query_get_parser)
+        @namespace.marshal_with(controller.flask_restx.get_response_model)
+        def get(self):
+            return []
+
+        @namespace.expect(controller.flask_restx.json_post_model)
+        def post(self):
+            return []
+
+        @namespace.expect(controller.flask_restx.json_put_model)
+        def put(self):
+            return []
+
+        @namespace.expect(controller.flask_restx.query_delete_parser)
+        def delete(self):
+            return []
+
+    return application
+
+
+def test_open_api_definition(client):
+    response = client.get("/swagger.json")
+    assert response.json == {
+        "swagger": "2.0",
+        "basePath": "/",
+        "paths": {
+            "/test": {
+                "post": {
+                    "responses": {"200": {"description": "Success"}},
+                    "operationId": "post_test_resource",
+                    "parameters": [
+                        {
+                            "name": "payload",
+                            "required": True,
+                            "in": "body",
+                            "schema": {
+                                "$ref": "#/definitions/TestCollection_PostRequestModel"
+                            },
+                        }
+                    ],
+                    "tags": ["Test"],
+                },
+                "put": {
+                    "responses": {"200": {"description": "Success"}},
+                    "operationId": "put_test_resource",
+                    "parameters": [
+                        {
+                            "name": "payload",
+                            "required": True,
+                            "in": "body",
+                            "schema": {
+                                "$ref": "#/definitions/TestCollection_PutRequestModel"
+                            },
+                        }
+                    ],
+                    "tags": ["Test"],
+                },
+                "delete": {
+                    "responses": {"200": {"description": "Success"}},
+                    "operationId": "delete_test_resource",
+                    "parameters": [
+                        {
+                            "name": "key",
+                            "in": "query",
+                            "type": "array",
+                            "items": {"type": "integer"},
+                            "collectionFormat": "multi",
+                        },
+                        {
+                            "name": "enum_field",
+                            "in": "query",
+                            "type": "array",
+                            "items": {"type": "string"},
+                            "collectionFormat": "multi",
+                        },
+                        {
+                            "name": "optional_with_default",
+                            "in": "query",
+                            "type": "array",
+                            "items": {"type": "string"},
+                            "collectionFormat": "multi",
+                        },
+                    ],
+                    "tags": ["Test"],
+                },
+                "get": {
+                    "responses": {
+                        "200": {
+                            "description": "Success",
+                            "schema": {
+                                "$ref": "#/definitions/TestCollection_GetResponseModel"
+                            },
+                        }
+                    },
+                    "operationId": "get_test_resource",
+                    "parameters": [
+                        {
+                            "name": "key",
+                            "in": "query",
+                            "type": "array",
+                            "items": {"type": "integer"},
+                            "collectionFormat": "multi",
+                        },
+                        {
+                            "name": "enum_field",
+                            "in": "query",
+                            "type": "array",
+                            "items": {"type": "string"},
+                            "collectionFormat": "multi",
+                        },
+                        {
+                            "name": "optional_with_default",
+                            "in": "query",
+                            "type": "array",
+                            "items": {"type": "string"},
+                            "collectionFormat": "multi",
+                        },
+                        {
+                            "name": "limit",
+                            "in": "query",
+                            "type": "integer",
+                            "minimum": 0,
+                            "exclusiveMinimum": True,
+                        },
+                        {
+                            "name": "offset",
+                            "in": "query",
+                            "type": "integer",
+                            "minimum": 0,
+                        },
+                        {
+                            "name": "X-Fields",
+                            "in": "header",
+                            "type": "string",
+                            "format": "mask",
+                            "description": "An optional fields mask",
+                        },
+                    ],
+                    "tags": ["Test"],
+                },
+            }
+        },
+        "info": {"title": "API", "version": "1.0"},
+        "produces": ["application/json"],
+        "consumes": ["application/json"],
+        "tags": [{"name": "Test"}],
+        "definitions": {
+            "TestCollection_PostRequestModel": {
+                "properties": {
+                    "enum_field": {
+                        "type": "string",
+                        "description": "Test Documentation",
+                        "readOnly": False,
+                        "example": "Value1",
+                        "enum": ["Value1", "Value2"],
+                    },
+                    "key": {"type": "integer", "readOnly": True, "example": 1},
+                    "optional_with_default": {
+                        "type": "string",
+                        "readOnly": False,
+                        "default": "Test value",
+                        "example": "Test value",
+                    },
+                },
+                "type": "object",
+            },
+            "TestCollection_PutRequestModel": {
+                "properties": {
+                    "enum_field": {
+                        "type": "string",
+                        "description": "Test Documentation",
+                        "readOnly": False,
+                        "example": "Value1",
+                        "enum": ["Value1", "Value2"],
+                    },
+                    "key": {"type": "integer", "readOnly": True, "example": 1},
+                    "optional_with_default": {
+                        "type": "string",
+                        "readOnly": False,
+                        "default": "Test value",
+                        "example": "Test value",
+                    },
+                },
+                "type": "object",
+            },
+            "TestCollection_GetResponseModel": {
+                "properties": {
+                    "enum_field": {
+                        "type": "string",
+                        "description": "Test Documentation",
+                        "readOnly": False,
+                        "example": "Value1",
+                        "enum": ["Value1", "Value2"],
+                    },
+                    "key": {"type": "integer", "readOnly": True, "example": 1},
+                    "optional_with_default": {
+                        "type": "string",
+                        "readOnly": False,
+                        "default": "Test value",
+                        "example": "Test value",
+                    },
+                },
+                "type": "object",
+            },
+        },
+        "responses": {
+            "ParseError": {"description": "When a mask can't be parsed"},
+            "MaskError": {"description": "When any error occurs on mask"},
+        },
+    }
